+++ nym-vpn-app/src/i18n/en/add-credential.json
@@ -1,13 +1,13 @@
 {
-  "welcome": "Welcome!",
-  "description": "Please enter your anonymous passphrase.",
-  "input-label": "Passphrase",
-  "input-placeholder": "24-word-passphrase\n\ne.g. smoke artefact velvet skull pop palace tortoise damage rough…",
-  "login-button": "Next",
+  "title": "Log in",
+  "description": "Enter or paste your passphrase: 24 words separated by spaces",
+  "input-label": "24 word-passphrase",
+  "input-placeholder": "word1 word2 word3 word4... (24 words)",
+  "login-button": "Log in",
   "error": "Invalid passphrase",
   "create-account": {
     "text": "New to NymVPN?",
-    "link": "Create an account"
+    "link": "Create account"
   },
   "added-notification": "Device successfully added to your account"
 }
