--- nym-vpn-app/src/screens/login/PassphraseLogin.tsx
@@ -1,7 +1,0 @@
-import OGLogin from './OGLogin';
-
-function PassphraseLogin() {
-  return <OGLogin />;
-}
-
-export default PassphraseLogin;
